RgbRenderer to separate module
