+++ src/renderer/mod.rs
@@ -1,37 +1,15 @@
+pub use rgb_renderer::RgbRenderer;
+
+use crate::hittable::Hittable;
 use crate::ray::Ray;
 use crate::vec::V3;
-use crate::hittable::Hittable;
+
+mod rgb_renderer;
 
 pub trait Renderer {
     fn color(&self, r: &Ray) -> V3;
 }
 
-pub struct RgbRenderer {
-    pub hittable: Box<dyn Hittable>
-}
-
-impl Renderer for RgbRenderer {
-    fn color(&self, r: &Ray) -> V3 {
-        match self.hittable.hit(&r, 0.0001, 99999.0) {
-            Some(hit) => {
-                let emitted = hit.material.emmit(&hit);
-                return match hit
-                    .material
-                    .scatter(r, &hit)
-                    .and_then(Ray::validate) {
-                    Some(scattered) => { emitted.0 + scattered.attenuation * self.color(&scattered) }
-                    None => emitted.0
-                };
-            }
-            None => {
-//                let unit_direction = r.direction.unit();
-//                let t: f64 = 0.5 * (unit_direction.y + 1.0);
-                return V3::new(0.05088, 0.05088, 0.05088);
-//                return (1.0 - t) * V3::ones() + t * V3::new(0.5, 0.7, 1.0);
-            }
-        };
-    }
-}
 
 pub struct TtlRenderer{
     hittable: Box<dyn Hittable>,
